refactor(app): update component styles, config and routes
